:sparkles: split ratls_certool subcommands
